merge

+++ JUmlIt/src/main/java/ua/nure/sigma/controller/DiagramController.java
@@ -5,6 +5,8 @@ import java.security.Principal;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,23 +17,25 @@ import org.springframework.web.bind.annotation.RestController;
 
 import ua.nure.sigma.db_entities.Diagram;
 import ua.nure.sigma.db_entities.User;
+import ua.nure.sigma.db_entities.diagram.Field;
 import ua.nure.sigma.exceptions.DiagramException;
 import ua.nure.sigma.service.AccountService;
 import ua.nure.sigma.service.DiagramService;
 import ua.nure.sigma.service.HistoryService;
+import ua.nure.sigma.util.UserAccessibility;
 
 @RestController
 public class DiagramController {
 
-	private final HttpSession session;
+	private final HttpSession httpSession;
 	private final SimpMessagingTemplate template;
 	private final HistoryService historyService;
 	private final DiagramService diagramService;
 	private final AccountService accountService;
-	
+
 	@Autowired
 	public DiagramController(final HttpSession session, final SimpMessagingTemplate template) {
-		this.session = session;
+		this.httpSession = session;
 		this.template = template;
 		this.historyService = new HistoryService(template);
 		this.accountService = new AccountService();
@@ -43,7 +47,6 @@ public class DiagramController {
 		User user = accountService.getUserByLogin(principal.getName());
 		diagram.setOwner(user);
 		Diagram diagramModel = diagramService.createDiagram(diagram);
-		historyService.insertHistory(principal, diagramModel, "inserted diagram");
 		if (diagramModel == null) {
 			throw new DiagramException("Cannot create diagram");
 		}
@@ -51,24 +54,38 @@ public class DiagramController {
 	}
 
 	@RequestMapping(value = "/diagram/{id}", method = RequestMethod.GET)
-	public Diagram getDiagramById(@PathVariable long id) {
+	public ResponseEntity<Diagram> getDiagramById(@PathVariable long id, Principal principal) {
+		if (!UserAccessibility.hasAccess(principal, id)) {
+			return new ResponseEntity<Diagram>(HttpStatus.FORBIDDEN);
+		}
 		Diagram model = diagramService.getDiagramById(id);
 		if (model == null) {
 			throw new DiagramException("Cannot retrieve diagram with id " + id);
 		}
-		return model;
+		return new ResponseEntity<Diagram>(model, HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/diagram/update", method = RequestMethod.POST)
-	public Diagram updateDiagram(@RequestBody Diagram diagram, Principal principal) {
+	public ResponseEntity<Diagram> updateDiagram(@RequestBody Diagram diagram, Principal principal) {
+		if (!UserAccessibility.hasAccess(principal, diagram.getDiagramId())) {
+			return new ResponseEntity<Diagram>(HttpStatus.FORBIDDEN);
+		}
 		diagramService.updateDiagram(diagram);
 		Diagram newDiagram = diagramService.getDiagramById(diagram.getDiagramId());
-		historyService.insertHistory(principal, newDiagram, "updated diagram");
-		return newDiagram;
+		historyService.insertHistory("updated diagram: " + diagram.getName(),
+				(Long) (httpSession.getAttribute("sessionId")));
+		return new ResponseEntity<Diagram>(newDiagram, HttpStatus.OK);
 	}
 
 	@RequestMapping(value = "/diagram/{id}/remove", method = RequestMethod.POST)
-	public void updateDiagram(@PathVariable long id) {
+	public ResponseEntity<Void> deleteDiagram(@PathVariable long id, Principal principal) {
+		if (!UserAccessibility.hasAccess(principal, id)) {
+			return new ResponseEntity<Void>(HttpStatus.FORBIDDEN);
+		}
+		Diagram diagram = diagramService.getDiagramById(id);
+		historyService.insertHistory("deleted diagram: " + diagram.getName(),
+				(Long) (httpSession.getAttribute("sessionId")));
 		diagramService.deleteDiagram(id);
+		return new ResponseEntity<Void>(HttpStatus.OK);
 	}
 }
